refactor: util 패키지 제거 및 하위 클래스들을 model로 이동
검증, 상수 클래스가 util 패키지에 묶여있을 필요가 없다고 판단하여 수정
