+++ src/main/java/baseball/model/Constant.java
@@ -1,4 +1,4 @@
-package baseball.model.util;
+package baseball.model;
 
 public interface Constant {
     String GAME_START_MESSAGE = "숫자 야구 게임을 시작합니다.";
