+++ src/main/java/baseball/model/Utilization.java
@@ -1,4 +1,4 @@
-package baseball.model.util;
+package baseball.model;
 
 import java.util.ArrayList;
 import java.util.Arrays;
