+++ src/main/java/baseball/model/Validation.java
@@ -1,4 +1,4 @@
-package baseball.model.util;
+package baseball.model;
 
 import java.util.List;
